back/requirements.txt
bcrypt==2.0.0
cffi==1.1.2
Flask==0.10.1
Flask-Classy==0.6.10
Flask-Cors==2.0.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
needs==1.0.9
pycparser==2.14
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
SQLAlchemy==1.0.7
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
